+++ packages/graphql-codegen-cli/src/cli.ts
@@ -1,22 +1,27 @@
 import * as commander from 'commander';
+import * as path from 'path';
+import * as fs from 'fs';
+import * as mkdirp from 'mkdirp';
+import * as validUrl from 'valid-url';
+
 import { introspectionFromFile } from './loaders/introspection-from-file';
 import { introspectionFromUrl } from './loaders/introspection-from-url';
 import { schemaFromExport } from './loaders/schema-from-export';
 import { documentsFromGlobs } from './utils/documents-glob';
-import { compileTemplate, FileOutput, ALLOWED_CUSTOM_TEMPLATE_EXT } from 'graphql-codegen-compiler';
+import { loadDocumentsSources } from './loaders/document-loader';
+import { scanForTemplatesInPath } from './loaders/templates-scanner';
+import { ALLOWED_CUSTOM_TEMPLATE_EXT, compileTemplate } from 'graphql-codegen-compiler';
 import {
   debugLog,
+  EInputType,
+  GeneratorConfig,
   introspectionToGraphQLSchema,
   schemaToTemplateContext,
-  transformDocument
+  transformDocument,
+  CustomProcessingFunction,
+  FileOutput,
+  isGeneratorConfig
 } from 'graphql-codegen-core';
-import { loadDocumentsSources } from './loaders/document-loader';
-import * as path from 'path';
-import * as fs from 'fs';
-import { scanForTemplatesInPath } from './loaders/templates-scanner';
-import { EInputType, GeneratorConfig, getGeneratorConfig } from 'graphql-codegen-generators';
-import * as mkdirp from 'mkdirp';
-import * as validUrl from 'valid-url';
 
 process.env.NODE_TLS_REJECT_UNAUTHORIZED = '0';
 
@@ -137,7 +142,7 @@ export const executeWithOptions = async (options: CLIOptions): Promise<FileOutpu
   const url: string = options.url;
   const fsExport: string = options.export;
   const documents: string[] = options.args || [];
-  const template: string = options.template;
+  let template: string = options.template;
   const project: string = options.project;
   const gqlGenConfigFilePath: string = options.config || './gql-gen.json';
   const out: string = options.out || './';
@@ -195,20 +200,31 @@ export const executeWithOptions = async (options: CLIOptions): Promise<FileOutpu
     graphQlSchema,
     loadDocumentsSources(await documentsFromGlobs(documents))
   );
-  let templateConfig: GeneratorConfig = null;
+  let templateConfig: GeneratorConfig | CustomProcessingFunction | null = null;
 
   if (template && template !== '') {
     debugLog(`[executeWithOptions] using template: ${template}`);
-    templateConfig = getGeneratorConfig(template);
 
-    if (!templateConfig) {
-      const templateFromExport = require(template);
+    // Backward compatibility for older versions
+    if (
+      template === 'ts' ||
+      template === 'ts-single' ||
+      template === 'typescript' ||
+      template === 'typescript-single'
+    ) {
+      template = 'graphql-codegen-typescript-template';
+    } else if (template === 'ts-multiple' || template === 'typescript-multiple') {
+      template = 'graphql-codegen-typescript-template-multiple';
+    }
 
-      if (!templateFromExport || !templateFromExport.default) {
-        throw new Error(`Unknown template: ${template}`);
-      } else {
-        templateConfig = templateFromExport.default;
-      }
+    const localFilePath = path.resolve(process.cwd(), template);
+    const localFileExists = fs.existsSync(localFilePath);
+    const templateFromExport = require(localFileExists ? localFilePath : template);
+
+    if (!templateFromExport) {
+      throw new Error(`Unknown codegen template: ${template}, please make sure it's installed using npm/Yarn!`);
+    } else {
+      templateConfig = templateFromExport.default || templateFromExport.config || templateFromExport;
     }
   }
 
@@ -259,12 +275,14 @@ export const executeWithOptions = async (options: CLIOptions): Promise<FileOutpu
     };
   }
 
-  templateConfig.config = config ? config.generatorConfig || {} : {};
+  if (isGeneratorConfig(templateConfig)) {
+    templateConfig.config = config ? config.generatorConfig || {} : {};
+  }
 
-  return compileTemplate(templateConfig, context, [transformedDocuments], {
+  return (await compileTemplate(templateConfig, context, [transformedDocuments], {
     generateSchema,
     generateDocuments
-  }).map((item: FileOutput) => {
+  })).map((item: FileOutput) => {
     let resultName = item.filename;
 
     if (!path.isAbsolute(resultName)) {