+++ packages/graphql-codegen-cli/tests/test-files/processor/valid.js
@@ -1,0 +1,3 @@
+module.exports = function(context, documents, settings) {
+  return [{ filename: 'a.ts', content: '1' }];
+};