+++ packages/graphql-codegen-compiler/src/generate-single-file.ts
@@ -1,7 +1,6 @@
 import { prepareSchemaForDocumentsOnly } from './prepare-documents-only';
-import { FileOutput, Settings } from './types';
-import { SchemaTemplateContext, Document, debugLog } from 'graphql-codegen-core';
-import { GeneratorConfig } from 'graphql-codegen-generators';
+import { Settings, FileOutput, SchemaTemplateContext, Document, debugLog } from 'graphql-codegen-core';
+import { GeneratorConfig } from 'graphql-codegen-core';
 import * as moment from 'moment';
 
 export function generateSingleFile(