+++ packages/templates/typescript-mongodb/src/helpers/if-not-root-type.ts
@@ -1,0 +1,9 @@
+function ifNotRootType(type, options) {
+  if (type.name !== 'Query' && type.name !== 'Mutation' && type.name !== 'Subscription') {
+    return options.fn(this);
+  } else {
+    return options.inverse(this);
+  }
+}
+
+export default ifNotRootType;